+++ Runtime/Scripts/Schema/BufferView.cs
@@ -6,12 +6,12 @@ using System;
 namespace GLTFast.Schema
 {
 
-    // public enum BufferViewTarget
-    // {
-    //     None = 0,
-    //     ArrayBuffer = 34962,
-    //     ElementArrayBuffer = 34963,
-    // }
+    public enum BufferViewTarget
+    {
+        None = 0,
+        ArrayBuffer = 34962,
+        ElementArrayBuffer = 34963,
+    }
 
     /// <inheritdoc/>
     [Serializable]