+++ Tests/Runtime/Scripts/Export/ExportTests.cs
@@ -117,9 +117,13 @@ namespace GLTFast.Tests.Export
             return guid;
         }
 
-        internal static void AddExportTestScene()
+        static void AddExportTestScene()
         {
-            var sceneName = GetExportSceneName();
+            AddExportTestScene(GetExportSceneName());
+        }
+
+        internal static void AddExportTestScene(string sceneName)
+        {
             var scenePath = $"Packages/{GltfGlobals.GltfPackageName}{k_ScenesPath}{sceneName}.unity";
             var sceneGuid = TryFixPackageAssetPath(ref scenePath);
 
@@ -135,7 +139,6 @@ namespace GLTFast.Tests.Export
             Array.Resize(ref scenes, scenes.Length+1);
             scenes[scenes.Length - 1] = new EditorBuildSettingsScene(sceneGuid, true);
             EditorBuildSettings.scenes = scenes;
-
         }
 
         static void CreateExportSceneObjectList(string sceneName)
@@ -548,7 +551,7 @@ namespace GLTFast.Tests.Export
                 "LayerMaskOne",
                 gameObjectExportSettings: new GameObjectExportSettings
                 {
-                    LayerMask = 1
+                    LayerMask = 2 // layer one: 1<<1
                 });
             yield return AsyncWrapper.WaitForTask(task);
 
@@ -558,7 +561,7 @@ namespace GLTFast.Tests.Export
                 "LayerMaskTwo",
                 gameObjectExportSettings: new GameObjectExportSettings
                 {
-                    LayerMask = 2
+                    LayerMask = 4 // layer two: 1<<2
                 });
             yield return AsyncWrapper.WaitForTask(task);